Address PR 4 review comments
- Add analyticsEvent to PaymentFlowAction so success button tracking
  is configurable per flow (cart tracks, bookings does not)
- Rename onBarcodeScanned to onSuccessScreenBarcodeScanned for clarity
- Split TotalsViewHelper: shared methods move to POSPaymentViewHelper,
  cart-specific methods stay in TotalsViewHelper for testability
