+++ Modules/Sources/PointOfSale/ViewHelpers/TotalsViewHelper.swift
@@ -1,104 +1,8 @@
 import Foundation
-import SwiftUI
-import WooFoundation
 
+/// Cart-specific view helper methods used only by TotalsView.
 struct TotalsViewHelper {
-    /// Shared background color for payment views (cart and bookings).
-    func paymentBackgroundColor(for paymentState: PointOfSalePaymentState) -> Color {
-        switch paymentState.activePaymentMethod {
-        case .cash:
-            switch paymentState.cash {
-            case .collectingCash, .paymentSuccess:
-                return .posSurfaceBright
-            default:
-                return .clear
-            }
-        case .card:
-            switch paymentState.card {
-            case .processingPayment:
-                return .posPrimary
-            case .cardPaymentSuccessful:
-                return .posSurfaceBright
-            default:
-                return .clear
-            }
-        }
-    }
-
-    /// Cash payment button visibility. Shows the button for zero-total orders (both cart and bookings)
-    /// and during card payment states where cash is a valid alternative.
-    func shouldShowCashPaymentButton(paymentState: PointOfSalePaymentState,
-                                     cardReaderConnectionStatus: CardPresentPaymentReaderConnectionStatus,
-                                     isZeroTotal: Bool = false) -> Bool {
-        if isZeroTotal, case .card = paymentState.activePaymentMethod {
-            return true
-        }
-        switch paymentState.activePaymentMethod {
-        case .cash:
-            return false
-        case .card:
-            if case .disconnected = cardReaderConnectionStatus,
-               case .idle = paymentState.card {
-                return true
-            }
-            switch paymentState.card {
-            case .validatingOrderError,
-                    .paymentIntentCreationError,
-                    .acceptingCard:
-                return true
-            default:
-                return false
-            }
-        }
-    }
-    func shouldShowTotalsFields(for paymentState: PointOfSalePaymentState) -> Bool {
-        switch paymentState.activePaymentMethod {
-        case .cash:
-            return false
-        case .card:
-            switch paymentState.card {
-            case .idle,
-                    .acceptingCard,
-                    .cardInserted,
-                    .validatingOrder,
-                    .validatingOrderError,
-                    .paymentIntentCreationError,
-                    .preparingReader:
-                return true
-            case .processingPayment,
-                    .paymentError,
-                    .cardPaymentSuccessful:
-                return false
-            }
-        }
-    }
-
-    func shouldShowDisconnectedMessage(readerConnectionStatus: CardPresentPaymentReaderConnectionStatus,
-                                       paymentState: PointOfSalePaymentState) -> Bool {
-        guard readerConnectionStatus == .disconnected else {
-            return false
-        }
-
-        switch paymentState.activePaymentMethod {
-        case .cash:
-            return false
-        case .card:
-            switch paymentState.card {
-            case .idle,
-                    .acceptingCard,
-                    .preparingReader:
-                return true
-            case .validatingOrder,
-                    .validatingOrderError,
-                    .paymentIntentCreationError,
-                    .processingPayment,
-                    .cardInserted,
-                    .paymentError,
-                    .cardPaymentSuccessful:
-                return false
-            }
-        }
-    }
+    private let paymentViewHelper = POSPaymentViewHelper()
 
     /// Cash payment button visibility for the cart flow (adds order state guards on top of the base check).
     func shouldShowCollectCashPaymentButton(orderState: PointOfSaleOrderState,
@@ -114,30 +18,9 @@ struct TotalsViewHelper {
             false
         }
 
-        return shouldShowCashPaymentButton(paymentState: paymentState,
-                                           cardReaderConnectionStatus: cardReaderConnectionStatus,
-                                           isZeroTotal: isZeroTotal)
-    }
-
-    func shouldApplyPadding(paymentState: PointOfSalePaymentState) -> Bool {
-        switch paymentState.activePaymentMethod {
-        case .cash:
-            return false
-        case .card:
-            switch paymentState.card {
-            case .cardPaymentSuccessful, .paymentError:
-                return false
-            case .idle,
-                    .acceptingCard,
-                    .cardInserted,
-                    .validatingOrder,
-                    .validatingOrderError,
-                    .paymentIntentCreationError,
-                    .preparingReader,
-                    .processingPayment:
-                return true
-            }
-        }
+        return paymentViewHelper.shouldShowCashPaymentButton(paymentState: paymentState,
+                                                             cardReaderConnectionStatus: cardReaderConnectionStatus,
+                                                             isZeroTotal: isZeroTotal)
     }
 
     func shouldShowTotalDiscountField(cart: Cart, orderTotals: PointOfSaleOrderTotals?) -> Bool {
